LayoutTests/http/tests/site-isolation/window-properties.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<iframe src="http://localhost:8000/site-isolation/resources/window-properties-child.html"></iframe>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}
addEventListener("message", (e) => {
    alert("postMessage received: " + e.data);
    if (window.testRunner) { testRunner.notifyDone() }
});
window.customFunction = function() {
    alert("custom function called");
}
</script>
