LayoutTests/http/tests/site-isolation/window-properties-expected.txt

ALERT: custom function called
ALERT: postMessage received: top children: 1, parent children: 1, parent parent is top: true
